Added compound examples
